minishell 20.05.19 1 leak

+++ srcs/hash_table/hash_table3.c
@@ -6,7 +6,7 @@
 /*   By: filip <filip@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/20 14:09:54 by filip             #+#    #+#             */
-/*   Updated: 2019/05/21 01:08:43 by filip            ###   ########.fr       */
+/*   Updated: 2019/05/21 12:37:10 by filip            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	free_hash_table(void)
 {
-	int		i;
+	short	i;
 	t_hash	**table;
 	t_hash	*next;
     t_hash  *copy;
@@ -30,7 +30,7 @@ void	free_hash_table(void)
             ft_memdel((void**)&(next->path));
             copy = next;
             next = next->next;
-            ft_memdel((void**)&(copy));
+			ft_memdel((void**)&(copy));
         }
 	}
 	ft_memdel((void**)table);
